+++ plugins_common.c
@@ -17,7 +17,7 @@
  * with this program; if not, write to the Free Software Foundation, Inc.,
  * 675 Mass Ave, Cambridge, MA 02139, USA.
  * 
- * $Id: plugins_common.c,v 1.1 2004/03/18 19:55:18 cheusov Exp $
+ * $Id: plugins_common.c,v 1.2 2004/11/17 12:39:45 cheusov Exp $
  * 
  */
 
@@ -57,7 +57,7 @@ static const int static_minus1_array [] = {
    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
 };
 
-const int *alloc_minus1_array (int count)
+const int *alloc_minus1_array (size_t count)
 {
    int *p;
